refactor(#414): extract DiffView from ChatConsole
